MWPW-174662: make a11y video controls bound to .hero-overlay
