+++ express/code/blocks/frictionless-quick-action/frictionless-quick-action.js
@@ -2,7 +2,6 @@ import { transformLinkToAnimation } from '../../scripts/utils/media.js';
 import { getLibs, getIconElementDeprecated, decorateButtonsDeprecated } from '../../scripts/utils.js';
 import { buildFreePlanWidget } from '../../scripts/widgets/free-plan.js';
 import { sendFrictionlessEventToAdobeAnaltics } from '../../scripts/instrument.js';
-import createAccessibilityVideoControls from '../../scripts/utils/ax-video-controls.js';
 
 let createTag; let getConfig;
 let getMetadata;
@@ -324,24 +323,13 @@ async function startSDKWithUnconvertedFile(file, quickAction, block) {
 }
 
 export default async function decorate(block) {
-  const [utils, gNavUtils, federated] = await Promise.all([import(`${getLibs()}/utils/utils.js`),
+  const [utils, gNavUtils] = await Promise.all([import(`${getLibs()}/utils/utils.js`),
     import(`${getLibs()}/blocks/global-navigation/utilities/utilities.js`),
-    import(`${getLibs()}/utils/federated.js`),
     decorateButtonsDeprecated(block)]);
 
   ({ createTag, getMetadata, loadScript, getConfig } = utils);
-  const { getFederatedContentRoot } = federated;
-
-  /** Localization for video labels */
-  const { replaceKeyArray } = await import(`${getLibs()}/features/placeholders.js`);
-  const [playAnimation, pauseAnimation] = await replaceKeyArray(['play-animation', 'pause-animation'], getConfig());
-  const videoLabels = {
-    playMotion: playAnimation || 'Play',
-    pauseMotion: pauseAnimation || 'Pause',
-  };
 
   globalNavSelector = gNavUtils?.selectors.globalNav;
-  const federatedRootPath = getFederatedContentRoot();
 
   const rows = Array.from(block.children);
   rows[1].classList.add('fqa-container');
@@ -363,13 +351,7 @@ export default async function decorate(block) {
   const dropzoneContainer = createTag('div', { class: 'dropzone-container' });
 
   if (animation && animation.href.includes('.mp4')) {
-    animationContainer.append(
-      createAccessibilityVideoControls(
-        transformLinkToAnimation(animation),
-        videoLabels,
-        federatedRootPath,
-      ),
-    );
+    animationContainer.append(transformLinkToAnimation(animation));
   }
 
   if (cta) cta.classList.add('xlarge');
